test_generator/requirements.txt
psycopg2-binary==2.9.5
Faker==13.15.0
